pyproject.toml
[project]
name = "deployaiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "langchain-agent>=0.1.1",
    "langchain-openai>=1.1.6",
    "numpy>=2.4.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
